requirements.txt
pymongo>=4.6.0
pydantic>=2.5.0
pyyaml>=6.0.1
click>=8.1.7
python-dotenv>=1.0.0
rich>=13.7.0
pytest>=8.0.0
fastapi>=0.110.0
uvicorn>=0.27.0
